refactor(watch): #1126 setup — split socket/runtime out of watch.rs
Pure carve-out, no logic changes. Precommit checkpoint before remaining
modules (rebuild, gc, events, reindex) split.

- src/cli/watch.rs (5711 lines) → src/cli/watch/mod.rs (5264 lines after socket+runtime)
- src/cli/watch/socket.rs (380 lines): SocketCleanupGuard, handle_socket_client,
  write_daemon_error{,_tracked}, trim_trailing_newline, max_concurrent_daemon_clients
- src/cli/watch/runtime.rs (90 lines): SHUTDOWN_REQUESTED, is_shutdown_requested,
  daemon_should_exit, on_sigterm, install_sigterm_handler, build_shared_runtime

Items reachable from cmd_watch are pub(super); the rest stays private. All 12
adversarial socket tests + the 5 trim_trailing_newline tests pass.
